Add delay between the connection of the servers
The Narupa server needs the VMD-IMD server of the OpenMM simulation to
accept connections. There may be a delay between the start of the OpenMM
process and the moment the server is ready. This delay leads to the
Narupa server trying to connect to early and failing. This commit
introduces a sleep between the start of the 2 server processes (OpenMM
VMD-IMD and Narupa) to avoid the failure.